refactor : repeat 함수 적용하기 위해 int형으로 변경

+++ src/main/java/racingcar/car/Car.java
@@ -5,7 +5,7 @@ import camp.nextstep.edu.missionutils.Randoms;
 public class Car {
 	private final String name;
 
-	private long location;
+	private int location;
 
 	public Car(String name) {
 		validateCarLength(name);
